api: simplify sync flow and centralize locking for user operations
- Move status management from router layer to separate state modules.
- Add to_dict serialization for EnsureUserStatus model.
- Update ensure_user_websocket to use status serialization.
- Wrap ensure_user in an async function with user-scoped lock.
